backend/requirements.txt
flask==2.3.3
flask-cors>=4.0.0
numpy<2.0.0
torch>=2.1.0
transformers>=4.46.0
requests>=2.25.0
pandas>=1.3.0
scikit-learn>=1.0.0
datasets>=2.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
rapidfuzz>=3.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-core>=0.2.26
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0 